// client-cargo-main (1)/client-cargo-main/src/stores/Trip/Trip.interface.ts
export enum TaxiClassEnum {
  ECONOMY = 'ECONOMY',
  COMFORT = 'COMFORT',
  PERSONAL = 'PERSONAL',
  BUSINESS = 'BUSINESS',
  TAXI = 'TAXI',
  CARSHARING = 'CARSHARING',
  BICYCLE = 'BICYCLE',
  WALK = 'WALK',
  PUBLIC = 'PUBLIC',
  SCOOTER = 'SCOOTER',
  YANDEX = 'YANDEX',
  CITYMOBIL = 'CITYMOBIL',
  UBER = 'UBER',

  VIP_BUS = 'VIP_BUS',
  SMALL_BUS = 'SMALL_BUS',
  MIDDLE_BUS = 'MIDDLE_BUS',
  LARGE_BUS = 'LARGE_BUS',
}
// ...
export const IOTripStops = t.intersection([
  t.type({
    orderId: t.number,
    address: t.string,
    latitude: t.number,
    longitude: t.number,
    startTime: t.union([t.number, t.string]),
    endTime: t.union([t.number, t.string]),
    eventType: t.string,
  }),
  t.partial({
    state: t.string,
  }),
]);
// ...
export const IOOrdersKpi = t.intersection([
  t.type({
    orderId: t.string,
    costSharePart: t.number,
    rideTimeMin: t.number,
    savings: t.number,
    savingsPct: t.number,
    orderDistanceKm: t.number,
  }),
  t.partial({
    state: t.string,
  }),
]);

export const IOTripKpi = t.intersection([
  t.type({
    totalCost: tt.money,
    totalDistanceKm: t.number,
    totalTimeMin: t.number,
    ordersKpi: t.array(IOOrdersKpi),
  }),
  t.partial({
    state: t.string,
  }),
]);

export const IOTripCoop = t.type({
  id: t.string,
  passengers: t.number,
  employeePassengers: t.array(Employee),
  newOrdersIds: t.array(t.string),
  stops: t.array(IOTripStops),
  kpi: IOTripKpi,
});

export type TTripStops = t.TypeOf<typeof IOTripStops>;
export type TTripFromCoopStops = t.TypeOf<typeof IOTripFromCoopStops>;
export type TTripKpi = t.TypeOf<typeof IOTripKpi>;
export type TTripCoop = t.TypeOf<typeof IOTripCoop>;